Contents of DataHaven Community Wellbeing Crosstabs

This is a nested data frame where each row corresponds to a year / set of years and location of the survey. The data column contains the full set of data extracted from the respective crosstabs; this, in turn, is nested by question code. This is a bit of a strange format, but it balances ease of subsetting by year and location with saving space and avoiding repeated values (e.g. listing out full question text dozens of times). Previously this data frame also included full text of every question, but as the size of the data has ballooned, these are now in the cws_codebook data frame, which can be joined to this by year and code.

Usage

cws_full_data

Format

A data frame of 385 rows and 4 columns.

Outer structure:

year

Numeric, year of survey

span

Character, span of years of the survey (e.g. "2015_2024")

name

Text of location

survey

List-column of data frames of survey response data. The number of rows varies based on the questions and participant groups available, but the 3 columns are the same.

For the survey list-column:

code

Question code, e.g. "Q2", "Q4E", "RENTEVICT"

data

List-column of more data frames, providing the actual response values per question. Again, number of rows varies, but the 4 columns are the same.

For the data list-column:

category

Factor: participant group categories, e.g. "Gender", "Age"

group

Factor: participant group, e.g. "Male", "Ages 65+"

response

Text of responses, depending on question

value

Share of participants giving each response

Source

Compiled DCWS crosstabs

Details

On its own, the structure is probably annoying. Easier extraction is available using fetch_cws.

See also

Examples

# get specific question based on code
cws_full_data |>
    dplyr::filter(span == "2015_2024", name == "Greater New Haven") |>
    tidyr::unnest(survey) |>
    dplyr::filter(code == "Q64") |>
    tidyr::unnest(data)
#> # A tibble: 76 × 8
#>     year span      name              code  category group       response   value
#>    <dbl> <chr>     <chr>             <chr> <fct>    <fct>       <fct>      <dbl>
#>  1  2024 2015_2024 Greater New Haven Q64   Total    Connecticu… Yes      0.100  
#>  2  2024 2015_2024 Greater New Haven Q64   Total    Connecticu… No       0.894  
#>  3  2024 2015_2024 Greater New Haven Q64   Total    Connecticu… Don't k… 0.00152
#>  4  2024 2015_2024 Greater New Haven Q64   Total    Connecticu… Refused  0.00410
#>  5  2024 2015_2024 Greater New Haven Q64   Total    Greater Ne… Yes      0.112  
#>  6  2024 2015_2024 Greater New Haven Q64   Total    Greater Ne… No       0.882  
#>  7  2024 2015_2024 Greater New Haven Q64   Total    Greater Ne… Don't k… 0.00175
#>  8  2024 2015_2024 Greater New Haven Q64   Total    Greater Ne… Refused  0.00395
#>  9  2024 2015_2024 Greater New Haven Q64   Gender   Male        Yes      0.112  
#> 10  2024 2015_2024 Greater New Haven Q64   Gender   Male        No       0.882  
#> # ℹ 66 more rows

# join first with codebook to find question by text
cws_full_data |>
    dplyr::filter(span == "2015_2024", name == "Greater New Haven") |>
    tidyr::unnest(survey) |>
    dplyr::left_join(cws_codebook, by = c("year", "code")) |>
    dplyr::filter(grepl("adequate shelter", question)) |>
    tidyr::unnest(data)
#> # A tibble: 76 × 10
#>     year span     name  code  category group response   value question responses
#>    <dbl> <chr>    <chr> <chr> <fct>    <fct> <fct>      <dbl> <chr>    <list>   
#>  1  2024 2015_20… Grea… Q64   Total    Conn… Yes      0.100   In the … <chr [4]>
#>  2  2024 2015_20… Grea… Q64   Total    Conn… No       0.894   In the … <chr [4]>
#>  3  2024 2015_20… Grea… Q64   Total    Conn… Don't k… 0.00152 In the … <chr [4]>
#>  4  2024 2015_20… Grea… Q64   Total    Conn… Refused  0.00410 In the … <chr [4]>
#>  5  2024 2015_20… Grea… Q64   Total    Grea… Yes      0.112   In the … <chr [4]>
#>  6  2024 2015_20… Grea… Q64   Total    Grea… No       0.882   In the … <chr [4]>
#>  7  2024 2015_20… Grea… Q64   Total    Grea… Don't k… 0.00175 In the … <chr [4]>
#>  8  2024 2015_20… Grea… Q64   Total    Grea… Refused  0.00395 In the … <chr [4]>
#>  9  2024 2015_20… Grea… Q64   Gender   Male  Yes      0.112   In the … <chr [4]>
#> 10  2024 2015_20… Grea… Q64   Gender   Male  No       0.882   In the … <chr [4]>
#> # ℹ 66 more rows

# make things easier with fetch_cws: flexibly grab by location, year, and/or
# filter conditions
fetch_cws(grepl("adequate shelter", question),
    .year = "2015_2024",
    .name = c("Connecticut", "Greater New Haven", "New Haven")
)
#> # A tibble: 3 × 6
#>    year span      name              code  question                      data    
#>   <dbl> <chr>     <chr>             <chr> <chr>                         <list>  
#> 1  2024 2015_2024 Connecticut       Q64   In the last 12 months, have … <tibble>
#> 2  2024 2015_2024 Greater New Haven Q64   In the last 12 months, have … <tibble>
#> 3  2024 2015_2024 New Haven         Q64   In the last 12 months, have … <tibble>