Contents of DataHaven Community Wellbeing Crosstabs

This is a nested data frame where each row corresponds to a year and location of the survey. The data column contains the full set of data extracted from the respective crosstabs; this, in turn, is nested by question and question code. This is a bit of a strange format, but it balances ease of subsetting by year and location with saving space and avoiding repeated values (e.g. listing out full question text dozens of times).

Usage

cws_full_data

Format

A data frame of 97 rows and 3 columns.

Outer structure:

year

Numeric, year of survey

name

Text of location

survey

List-column of data frames of survey response data. The number of rows varies based on the questions and participant groups available, but the 3 columns are the same.

For the survey list-column:

code

Question code, e.g. "Q2", "Q4E", "RENTEVICT"

question

Full text of survey question

data

List-column of more data frames, providing the actual response values per question. Again, number of rows varies, but the 4 columns are the same.

For the data list-column:

category

Factor: participant group categories, e.g. "Gender", "Age"

group

Factor: participant group, e.g. "Male", "Ages 65+"

response

Text of responses, depending on question

value

Share of participants giving each response

Source

Compiled DCWS crosstabs

Details

On its own, the structure is probably annoying. Easier extraction is available using fetch_cws.

See also

Examples

# bunch of different ways to work with this
# specific question for one location and one year
cws_full_data %>%
  dplyr::filter(year == 2018, name == "Greater New Haven") %>%
  tidyr::unnest(survey) %>%
  dplyr::filter(grepl("suitable employment", question)) %>%
  tidyr::unnest(data)
#> # A tibble: 138 × 8
#>     year name              code  question            categ…¹ group respo…² value
#>    <dbl> <chr>             <chr> <chr>               <fct>   <fct> <chr>   <dbl>
#>  1  2018 Greater New Haven Q4E   The ability of res… Total   Conn… Excell…  0.09
#>  2  2018 Greater New Haven Q4E   The ability of res… Total   Grea… Excell…  0.08
#>  3  2018 Greater New Haven Q4E   The ability of res… Gender  Male  Excell…  0.08
#>  4  2018 Greater New Haven Q4E   The ability of res… Gender  Fema… Excell…  0.08
#>  5  2018 Greater New Haven Q4E   The ability of res… Age     Ages… Excell…  0.11
#>  6  2018 Greater New Haven Q4E   The ability of res… Age     Ages… Excell…  0.09
#>  7  2018 Greater New Haven Q4E   The ability of res… Age     Ages… Excell…  0.06
#>  8  2018 Greater New Haven Q4E   The ability of res… Age     Ages… Excell…  0.06
#>  9  2018 Greater New Haven Q4E   The ability of res… Race/E… White Excell…  0.07
#> 10  2018 Greater New Haven Q4E   The ability of res… Race/E… Black Excell…  0.09
#> # … with 128 more rows, and abbreviated variable names ¹​category, ²​response

# simpler but a bit slower--has to unnest to full 600k rows
cws_full_data %>%
  tidyr::unnest(survey) %>%
  tidyr::unnest(data) %>%
  dplyr::filter(year == 2018,
                name == "Greater New Haven",
                grepl("suitable employment", question))
#> # A tibble: 138 × 8
#>     year name              code  question            categ…¹ group respo…² value
#>    <dbl> <chr>             <chr> <chr>               <fct>   <fct> <chr>   <dbl>
#>  1  2018 Greater New Haven Q4E   The ability of res… Total   Conn… Excell…  0.09
#>  2  2018 Greater New Haven Q4E   The ability of res… Total   Grea… Excell…  0.08
#>  3  2018 Greater New Haven Q4E   The ability of res… Gender  Male  Excell…  0.08
#>  4  2018 Greater New Haven Q4E   The ability of res… Gender  Fema… Excell…  0.08
#>  5  2018 Greater New Haven Q4E   The ability of res… Age     Ages… Excell…  0.11
#>  6  2018 Greater New Haven Q4E   The ability of res… Age     Ages… Excell…  0.09
#>  7  2018 Greater New Haven Q4E   The ability of res… Age     Ages… Excell…  0.06
#>  8  2018 Greater New Haven Q4E   The ability of res… Age     Ages… Excell…  0.06
#>  9  2018 Greater New Haven Q4E   The ability of res… Race/E… White Excell…  0.07
#> 10  2018 Greater New Haven Q4E   The ability of res… Race/E… Black Excell…  0.09
#> # … with 128 more rows, and abbreviated variable names ¹​category, ²​response

# specific question, one location, multiple years
cws_full_data %>%
  dplyr::mutate(survey = purrr::map(survey, dplyr::filter, question == "Diabetes")) %>%
  dplyr::filter(name == "New Haven") %>%
  tidyr::unnest(survey) %>%
  tidyr::unnest(data)
#> # A tibble: 236 × 8
#>     year name      code  question category       group       response value
#>    <dbl> <chr>     <chr> <chr>    <fct>          <fct>       <chr>    <dbl>
#>  1  2015 New Haven Q23C  Diabetes Total          Connecticut Yes       0.09
#>  2  2015 New Haven Q23C  Diabetes Total          New Haven   Yes       0.1 
#>  3  2015 New Haven Q23C  Diabetes Gender         Male        Yes       0.09
#>  4  2015 New Haven Q23C  Diabetes Gender         Female      Yes       0.12
#>  5  2015 New Haven Q23C  Diabetes Age            Ages 18-34  Yes       0.01
#>  6  2015 New Haven Q23C  Diabetes Age            Ages 35-49  Yes       0.08
#>  7  2015 New Haven Q23C  Diabetes Age            Ages 50-64  Yes       0.27
#>  8  2015 New Haven Q23C  Diabetes Age            Ages 65+    Yes       0.24
#>  9  2015 New Haven Q23C  Diabetes Race/Ethnicity White       Yes       0.07
#> 10  2015 New Haven Q23C  Diabetes Race/Ethnicity Black       Yes       0.15
#> # … with 226 more rows

# make things easier with fetch_cws: flexibly grab by location, year, and/or
# filter conditions
fetch_cws(grepl("Have there been times .+ food", question), .year = 2018,
          .name = c("Connecticut", "Greater New Haven", "New Haven"))
#> # A tibble: 3 × 5
#>    year name              code  question                                data    
#>   <dbl> <chr>             <chr> <chr>                                   <list>  
#> 1  2018 Connecticut       Q62   Have there been times in the past 12 m… <tibble>
#> 2  2018 Greater New Haven Q62   Have there been times in the past 12 m… <tibble>
#> 3  2018 New Haven         Q62   Have there been times in the past 12 m… <tibble>