Fetch and subset DCWS data

This function returns the doubly-nested data from cws_full_data in a nicer format, with options for subsetting. Filtering by year, location name, and category are named options, any of which take a vector of one or more values, but any valid conditions can be passed to ... for more flexible filtering. For any named options, NULL, the default, will mean no filtering is done by that column.

Usage

fetch_cws(
  ...,
  .year = NULL,
  .name = NULL,
  .category = NULL,
  .unnest = FALSE,
  .add_wts = FALSE
)

Arguments

...

Any number of conditions to filter by, which will be passed to dplyr::filter. These don't override the named options, so if you filter by year > 2020 but then set .year = 2015 you're not going to get any data.

.year

A vector of one or more numbers giving the year(s) to subset by. If NULL, no filtering is done by year.

.name

A vector of one or more strings giving the name(s) to subset by. If NULL, no filtering is done by name.

.category

A vector of one or more strings giving the category(ies) to subset by. If NULL, no filtering is done by category.

.unnest

Logical: should the data column be unnested? This just saves a step of calling tidyr::unnest but defaults to false.

.add_wts

Logical: should groups' survey weights be attached, via a left-join with dcws::cws_full_wts? This is useful if you need to collapse groups later; otherwise you might get stuck in annoying tidyr::unnest messes.

Value

A data frame, with between 5 and 9 columns, depending on arguments. Columns year, name, code, and question are always included. Additional columns:

arguments	columns

.unnest = F, .add_wts = F	data (nested df of category, group, response, and value)

.unnest = F, .add_wts = T	data (nested df of category, group, response, value, and weight)

.unnest = T, .add_wts = F	category, group, response, value

.unnest = T, .add_wts = T	category, group, response, value, weight

See also

Examples

# no filtering
fetch_cws()
#> # A tibble: 10,274 × 5
#>     year name  code  question                                        data       
#>    <dbl> <chr> <chr> <chr>                                           <list>     
#>  1  2015 5CT   Q1    Are you satisfied with the city or area where … <tibble [2…
#>  2  2015 5CT   Q2    As a place to live, is the city or area where … <tibble [4…
#>  3  2015 5CT   Q3A   How responsive local government is to the need… <tibble [3…
#>  4  2015 5CT   Q3B   The availability of the goods and services tha… <tibble [3…
#>  5  2015 5CT   Q3C   The job done by the police to keep residents s… <tibble [3…
#>  6  2015 5CT   Q3D   The ability of residents to obtain suitable em… <tibble [3…
#>  7  2015 5CT   Q3E   As a place to raise children                    <tibble [3…
#>  8  2015 5CT   Q3F   The condition of public parks and other public… <tibble [3…
#>  9  2015 5CT   Q4    Over the past 12 months, have you volunteered … <tibble [2…
#> 10  2015 5CT   Q5    How would you describe your ability to influen… <tibble [3…
#> # … with 10,264 more rows

# filter by year, name, and/or category
fetch_cws(.name = c("Greater New Haven", "New Haven")) # all years
#> # A tibble: 843 × 5
#>     year name              code  question                             data      
#>    <dbl> <chr>             <chr> <chr>                                <list>    
#>  1  2015 Greater New Haven Q1    Are you satisfied with the city or … <tibble […
#>  2  2015 Greater New Haven Q2    As a place to live, is the city or … <tibble […
#>  3  2015 Greater New Haven Q3A   How responsive local government is … <tibble […
#>  4  2015 Greater New Haven Q3B   The availability of the goods and s… <tibble […
#>  5  2015 Greater New Haven Q3C   The job done by the police to keep … <tibble […
#>  6  2015 Greater New Haven Q3D   The ability of residents to obtain … <tibble […
#>  7  2015 Greater New Haven Q3E   As a place to raise children         <tibble […
#>  8  2015 Greater New Haven Q3F   The condition of public parks and o… <tibble […
#>  9  2015 Greater New Haven Q4    Over the past 12 months, have you v… <tibble […
#> 10  2015 Greater New Haven Q5    How would you describe your ability… <tibble […
#> # … with 833 more rows
fetch_cws(.year = 2018, .name = c("Greater New Haven", "New Haven"))
#> # A tibble: 297 × 5
#>     year name              code  question                             data      
#>    <dbl> <chr>             <chr> <chr>                                <list>    
#>  1  2018 Greater New Haven Q1    Are you satisfied with the city or … <tibble […
#>  2  2018 Greater New Haven Q2    As a place to live, is the city or … <tibble […
#>  3  2018 Greater New Haven Q4A   How responsive local government is … <tibble […
#>  4  2018 Greater New Haven Q4D   The job done by the police to keep … <tibble […
#>  5  2018 Greater New Haven Q4E   The ability of residents to obtain … <tibble […
#>  6  2018 Greater New Haven Q4F   As a place to raise children         <tibble […
#>  7  2018 Greater New Haven Q4G   The condition of public parks and o… <tibble […
#>  8  2018 Greater New Haven Q4H   The availability of affordable, hig… <tibble […
#>  9  2018 Greater New Haven Q5    Over the past 12 months, have you v… <tibble […
#> 10  2018 Greater New Haven Q6    How would you describe your ability… <tibble […
#> # … with 287 more rows
fetch_cws(.year = 2021, .name = "New Haven", .category = c("Total", "Age", "Gender"))
#> # A tibble: 166 × 5
#>     year name      code   question                                    data      
#>    <dbl> <chr>     <chr>  <chr>                                       <list>    
#>  1  2021 New Haven Q1     Are you satisfied with the city or area wh… <tibble […
#>  2  2021 New Haven Q2     As a place to live, is the city or area wh… <tibble […
#>  3  2021 New Haven Q4A    How responsive local government is to the … <tibble […
#>  4  2021 New Haven Q4D    The job done by the police to keep residen… <tibble […
#>  5  2021 New Haven Q4E    The ability of residents to obtain suitabl… <tibble […
#>  6  2021 New Haven Q4F    As a place to raise children                <tibble […
#>  7  2021 New Haven Q4G    The condition of public parks and other pu… <tibble […
#>  8  2021 New Haven Q4H    The availability of affordable, high-quali… <tibble […
#>  9  2021 New Haven Q6     How would you describe your ability to inf… <tibble […
#> 10  2021 New Haven TRUSTA How much trust do you have in each of the … <tibble […
#> # … with 156 more rows

# filter by conditions
fetch_cws(code == "Q4E", .year = 2018, .name = c("Greater New Haven", "New Haven"), .unnest = TRUE)
#> # A tibble: 252 × 8
#>     year name              code  question        category  group  response value
#>    <dbl> <chr>             <chr> <chr>           <fct>     <fct>  <chr>    <dbl>
#>  1  2018 Greater New Haven Q4E   The ability of… Total     Conne… Excelle…  0.09
#>  2  2018 Greater New Haven Q4E   The ability of… Total     Great… Excelle…  0.08
#>  3  2018 Greater New Haven Q4E   The ability of… Gender    Male   Excelle…  0.08
#>  4  2018 Greater New Haven Q4E   The ability of… Gender    Female Excelle…  0.08
#>  5  2018 Greater New Haven Q4E   The ability of… Age       Ages … Excelle…  0.11
#>  6  2018 Greater New Haven Q4E   The ability of… Age       Ages … Excelle…  0.09
#>  7  2018 Greater New Haven Q4E   The ability of… Age       Ages … Excelle…  0.06
#>  8  2018 Greater New Haven Q4E   The ability of… Age       Ages … Excelle…  0.06
#>  9  2018 Greater New Haven Q4E   The ability of… Race/Eth… White  Excelle…  0.07
#> 10  2018 Greater New Haven Q4E   The ability of… Race/Eth… Black  Excelle…  0.09
#> # … with 242 more rows
fetch_cws(grepl("Q4[A-Z]", code), .year = 2018, .name = c("Greater New Haven", "New Haven"))
#> # A tibble: 12 × 5
#>     year name              code  question                             data      
#>    <dbl> <chr>             <chr> <chr>                                <list>    
#>  1  2018 Greater New Haven Q4A   How responsive local government is … <tibble […
#>  2  2018 Greater New Haven Q4D   The job done by the police to keep … <tibble […
#>  3  2018 Greater New Haven Q4E   The ability of residents to obtain … <tibble […
#>  4  2018 Greater New Haven Q4F   As a place to raise children         <tibble […
#>  5  2018 Greater New Haven Q4G   The condition of public parks and o… <tibble […
#>  6  2018 Greater New Haven Q4H   The availability of affordable, hig… <tibble […
#>  7  2018 New Haven         Q4A   How responsive local government is … <tibble […
#>  8  2018 New Haven         Q4D   The job done by the police to keep … <tibble […
#>  9  2018 New Haven         Q4E   The ability of residents to obtain … <tibble […
#> 10  2018 New Haven         Q4F   As a place to raise children         <tibble […
#> 11  2018 New Haven         Q4G   The condition of public parks and o… <tibble […
#> 12  2018 New Haven         Q4H   The availability of affordable, hig… <tibble […
fetch_cws(grepl("health insurance", question), year > 2015, .name = "New Haven")
#> # A tibble: 6 × 5
#>    year name      code  question                                    data        
#>   <dbl> <chr>     <chr> <chr>                                       <list>      
#> 1  2018 New Haven Q26   Do you have health insurance?               <tibble [76…
#> 2  2018 New Haven Q27   (If have health insurance) What type of he… <tibble [12…
#> 3  2018 New Haven Q30B  The doctor or hospital wouldn't accept you… <tibble [8 …
#> 4  2021 New Haven Q26   Do you have health insurance?               <tibble [76…
#> 5  2021 New Haven Q27   (If yes) What type of health insurance do … <tibble [16…
#> 6  2021 New Haven Q30B  The doctor or hospital wouldn't accept you… <tibble [4 …
fetch_cws(question %in% c("Diabetes", "Asthma"), .name = "Bridgeport")
#> # A tibble: 6 × 5
#>    year name       code  question data             
#>   <dbl> <chr>      <chr> <chr>    <list>           
#> 1  2015 Bridgeport Q23C  Diabetes <tibble [84 × 4]>
#> 2  2015 Bridgeport Q23E  Asthma   <tibble [84 × 4]>
#> 3  2018 Bridgeport Q23C  Diabetes <tibble [76 × 4]>
#> 4  2018 Bridgeport Q23E  Asthma   <tibble [76 × 4]>
#> 5  2021 Bridgeport Q23C  Diabetes <tibble [76 × 4]>
#> 6  2021 Bridgeport Q23E  Asthma   <tibble [76 × 4]>

# how you might use this to make a beautiful table
fetch_cws(code == "Q1", .year = 2021, .category = c("Income", "Gender"), .unnest = TRUE) %>%
  dplyr::group_by(name, category, group) %>%
  cwi::sub_nonanswers() %>%
  dplyr::filter(response == "Yes") %>%
  tidyr::pivot_wider(id_cols = name, names_from = group, values_from = value)
#> # A tibble: 19 × 6
#> # Groups:   name [19]
#>    name                    Male Female `<$30K` `$30K-$100K` `$100K+`
#>    <chr>                  <dbl>  <dbl>   <dbl>        <dbl>    <dbl>
#>  1 Bridgeport             0.788  0.778   0.758        0.778    0.814
#>  2 Connecticut            0.888  0.889   0.798        0.879    0.919
#>  3 Fairfield              0.88   0.88    0.69         0.79     0.96 
#>  4 Fairfield County       0.87   0.899   0.798        0.889    0.929
#>  5 Greater Hartford       0.898  0.89    0.828        0.908    0.960
#>  6 Greater New Haven      0.869  0.879   0.828        0.899    0.87 
#>  7 Greenwich              0.89   0.980   0.94         0.929    0.939
#>  8 Groton                 0.78   0.884   0.835        0.74     0.89 
#>  9 Hartford               0.716  0.75    0.673        0.760    0.818
#> 10 Lower Naugatuck Valley 0.86   0.929   0.85         0.888    0.95 
#> 11 Middlesex County       0.92   0.94    0.77         0.93     0.96 
#> 12 Milford                0.88   0.929   0.89         0.929    0.86 
#> 13 New Haven              0.818  0.838   0.765        0.828    0.929
#> 14 New London             0.908  0.859   0.88         0.939    0.72 
#> 15 Stamford               0.9    0.908   0.869        0.929    0.9  
#> 16 Stratford              0.84   0.928   0.74         0.918    0.89 
#> 17 Trumbull               0.87   0.949   0.85         0.94     0.85 
#> 18 Waterbury              0.74   0.670   0.55         0.737    0.81 
#> 19 West Haven             0.711  0.939   0.918        0.773    0.806

# adding weights to collapse groups (e.g. combining income brackets)
fetch_cws(code == "Q1", .year = 2021, .add_wts = TRUE)
#> # A tibble: 20 × 5
#>     year name                   code  question                        data      
#>    <dbl> <chr>                  <chr> <chr>                           <list>    
#>  1  2021 Bridgeport             Q1    Are you satisfied with the cit… <tibble […
#>  2  2021 Connecticut            Q1    Are you satisfied with the cit… <tibble […
#>  3  2021 Fairfield              Q1    Are you satisfied with the cit… <tibble […
#>  4  2021 Fairfield County       Q1    Are you satisfied with the cit… <tibble […
#>  5  2021 Greater Hartford       Q1    Are you satisfied with the cit… <tibble […
#>  6  2021 Greater New Haven      Q1    Are you satisfied with the cit… <tibble […
#>  7  2021 Greenwich              Q1    Are you satisfied with the cit… <tibble […
#>  8  2021 Groton                 Q1    Are you satisfied with the cit… <tibble […
#>  9  2021 Hartford               Q1    Are you satisfied with the cit… <tibble […
#> 10  2021 Lower Naugatuck Valley Q1    Are you satisfied with the cit… <tibble […
#> 11  2021 Middlesex County       Q1    Are you satisfied with the cit… <tibble […
#> 12  2021 Milford                Q1    Are you satisfied with the cit… <tibble […
#> 13  2021 Monroe                 Q1    Are you satisfied with the cit… <tibble […
#> 14  2021 New Haven              Q1    Are you satisfied with the cit… <tibble […
#> 15  2021 New London             Q1    Are you satisfied with the cit… <tibble […
#> 16  2021 Stamford               Q1    Are you satisfied with the cit… <tibble […
#> 17  2021 Stratford              Q1    Are you satisfied with the cit… <tibble […
#> 18  2021 Trumbull               Q1    Are you satisfied with the cit… <tibble […
#> 19  2021 Waterbury              Q1    Are you satisfied with the cit… <tibble […
#> 20  2021 West Haven             Q1    Are you satisfied with the cit… <tibble […
fetch_cws(.year = 2021, .name = "New Haven", .category = c("Total", "Age", "Income"),
          .add_wts = TRUE, .unnest = TRUE)
#> # A tibble: 4,205 × 9
#>     year name      code  question          category group  response value weight
#>    <dbl> <chr>     <chr> <chr>             <fct>    <fct>  <chr>    <dbl>  <dbl>
#>  1  2021 New Haven Q1    Are you satisfie… Total    Conne… Yes       0.87  1    
#>  2  2021 New Haven Q1    Are you satisfie… Total    New H… Yes       0.82  1    
#>  3  2021 New Haven Q1    Are you satisfie… Age      Ages … Yes       0.8   0.456
#>  4  2021 New Haven Q1    Are you satisfie… Age      Ages … Yes       0.82  0.219
#>  5  2021 New Haven Q1    Are you satisfie… Age      Ages … Yes       0.83  0.151
#>  6  2021 New Haven Q1    Are you satisfie… Age      Ages … Yes       0.88  0.173
#>  7  2021 New Haven Q1    Are you satisfie… Income   <$30K  Yes       0.75  0.325
#>  8  2021 New Haven Q1    Are you satisfie… Income   $30K-… Yes       0.82  0.515
#>  9  2021 New Haven Q1    Are you satisfie… Income   $100K+ Yes       0.92  0.16 
#> 10  2021 New Haven Q1    Are you satisfie… Total    Conne… No        0.12  1    
#> # … with 4,195 more rows